refactor(tests): split moderation routes for GH727
