+++ specs/GH727/tasks.md
@@ -282,24 +282,30 @@ GH-727 / #727
 - [x] `SP727-T269` Owner: coordinator. Done when: `src/core/providers/anthropic/client/tests.rs` becomes a small test facade with shared imports and child module declarations for setup/error, message/tool, response, and request edge tests. Verify: `cat src/core/providers/anthropic/client/tests.rs`.
 - [x] `SP727-T270` Owner: coordinator. Done when: original Anthropic client tests move into `src/core/providers/anthropic/client/tests/*.rs` by behavior domain without assertion changes. Verify: `rg -n "test_client_creation|test_header_building|test_map_http_error_400|test_extract_retry_after_from_root|test_separate_system_messages_no_system|test_anthropic_transform_messages_preserves_assistant_text_with_tool_use|test_transform_tool_choice_auto|test_transform_chat_response_text|test_anthropic_usage_cache_details|test_anthropic_client_preserves_thinking_blocks|test_transform_chat_request_rejects_n_greater_than_one|test_transform_chat_request_allows_unknown_model_when_configured|test_transform_chat_response_preserves_cache_details_without_double_counting" src/core/providers/anthropic/client/tests`.
 - [x] `SP727-T271` Owner: verification owner. Done when: all touched Anthropic client test files are below U-16's 800-line ceiling and focused Anthropic client tests pass. Verify: `wc -l src/core/providers/anthropic/client/tests.rs src/core/providers/anthropic/client/tests/*.rs`; `cargo test core::providers::anthropic::client::tests --lib --all-features` passed 39 tests.
-- [ ] `SP727-T272` Owner: verification owner. Done when: formatting, SpecRail, diff check, lib all-features check, locked all-features check, default check, PR CI, and review-thread gate pass for the Anthropic client tests tranche. Verify: `cargo fmt --all -- --check`; `git diff --check`; `python3 /Users/apple/Desktop/code/AI/tool/specrail/checks/check_workflow.py --repo /Users/apple/Desktop/code/AI/tool/specrail --spec-dir "$PWD/specs/GH727"`; `cargo check --lib --all-features`; `cargo check --all-features --locked`; `cargo check`; GitHub PR CI and review-thread query.
+- [x] `SP727-T272` Owner: verification owner. Done when: formatting, SpecRail, diff check, lib all-features check, locked all-features check, default check, PR CI, and review-thread gate pass for the Anthropic client tests tranche. Verify: #870 PR body, green PR CI, GraphQL reviewThreads totalCount 0, merge commit `68a17074c41c`, #727 reopened, and #727 update comment.
+- [x] `SP727-T273` Owner: coordinator. Done when: after the Anthropic client tests tranche merges, the next #727 tranche is selected from the remaining queue with fresh tracked-file line-count evidence. Verify: at `origin/main@68a17074`, `tests/moderations_routes.rs` is 809 lines and 6 tracked Rust files remain over the U-16 ceiling.
+- [x] `SP727-T274` Owner: coordinator. Done when: the moderation routes tranche documents why this slice is a test-suite split rather than a production route/router/auth/budget split. Verify: `git diff -- specs/GH727/product.md specs/GH727/tech.md`.
+- [x] `SP727-T275` Owner: coordinator. Done when: `tests/moderations_routes.rs` becomes a gated integration-test facade with shared mock upstream, app-state, provider, and auth helpers plus child module declarations for proxy/default-selection, auth/validation, and budget/fallback tests. Verify: `sed -n '1,260p' tests/moderations_routes.rs`.
+- [x] `SP727-T276` Owner: coordinator. Done when: original moderation route tests move into `tests/tests/moderations_routes_*.rs` by behavior domain without assertion changes. Verify: `rg -n "moderation_route_without_provider_fails_closed|moderation_route_proxies_request_with_provider_headers|root_moderation_alias_proxies_request|moderation_route_uses_default_model_for_provider_selection_when_omitted|moderation_route_requires_auth_when_anonymous_is_disabled|moderation_route_allows_authenticated_api_key|moderation_route_rejects_invalid_request_before_upstream|moderation_route_rejects_unconfigured_model_before_upstream|moderation_route_rejects_exhausted_provider_budget_before_upstream|moderation_route_uses_router_budget_fallback_provider|native_openai_moderation_route_uses_default_model_with_empty_config_models|openai_compatible_named_openai_uses_provider_name_wildcard_fallback|moderation_route_uses_wildcard_provider_name_fallback|moderation_route_rejects_exhausted_default_model_budget_before_upstream" tests/tests`.
+- [x] `SP727-T277` Owner: verification owner. Done when: all touched moderation route test files are below U-16's 800-line ceiling and focused moderation routes test passes. Verify: `wc -l tests/moderations_routes.rs tests/tests/moderations_routes_*.rs`; `cargo test --test moderations_routes --all-features` passed 14 tests.
+- [ ] `SP727-T278` Owner: verification owner. Done when: formatting, SpecRail, diff check, lib all-features check, locked all-features check, default check, PR CI, and review-thread gate pass for the moderation routes tranche. Verify: `cargo fmt --all -- --check`; `git diff --check`; `python3 /Users/apple/Desktop/code/AI/tool/specrail/checks/check_workflow.py --repo /Users/apple/Desktop/code/AI/tool/specrail --spec-dir "$PWD/specs/GH727"`; `cargo check --lib --all-features`; `cargo check --all-features --locked`; `cargo check`; GitHub PR CI and review-thread query.
 
 ## 并行拆分
 
-This is a serial writable lane for the Anthropic client tests file family. Other #727 large-file tranches may be planned read-only in parallel, but they must not edit this branch.
+This is a serial writable lane for the moderation routes test file family. Other #727 large-file tranches may be planned read-only in parallel, but they must not edit this branch.
 
 Writable ownership for this lane:
 
 - `specs/GH727/`
-- `src/core/providers/anthropic/client/tests.rs`
-- `src/core/providers/anthropic/client/tests/`
+- `tests/moderations_routes.rs`
+- `tests/tests/moderations_routes_*.rs`
 
 ## 验证
 
 - SpecRail packet review.
 - `cargo fmt --all -- --check`
 - `git diff --check`
-- `cargo test core::providers::anthropic::client::tests --lib --all-features`
+- `cargo test --test moderations_routes --all-features`
 - `cargo check --lib --all-features`
 - `cargo check --all-features --locked`
 - `cargo check`
@@ -307,5 +313,5 @@ Writable ownership for this lane:
 
 ## Handoff Notes
 
-This PR is the next #727 Anthropic client test-suite split tranche and should use `Refs #727`, not `Closes #727`.
+This PR is the next #727 moderation routes integration-test split tranche and should use `Refs #727`, not `Closes #727`.
 The issue should remain open until the final scan shows no Rust files over the U-16 ceiling.
